+++ verifier/src/test/java/com/aws/jverify/verifier/tests/generics/PolymorphismDafnyResolutionErrors2.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.tests;
+package com.aws.jverify.verifier.tests.generics;
 
 import com.aws.jverify.Modifiable;
 import com.aws.jverify.testengine.JVerifyTest;
