+++ verifier/src/test/java/com/aws/jverify/verifier/tests/generics/PolymorphismDafnyResolutionErrors3.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.tests;
+package com.aws.jverify.verifier.tests.generics;
 
 import com.aws.jverify.testengine.JVerifyTest;
