+++ verifier/src/test/java/com/aws/jverify/verifier/tests/generics/PolymorphismWithoutBounds.java
@@ -1,9 +1,7 @@
-package com.aws.jverify.verifier.tests;
+package com.aws.jverify.verifier.tests.generics;
 
 import com.aws.jverify.Pure;
 import com.aws.jverify.testengine.JVerifyTest;
-
-import javax.print.attribute.PrintJobAttribute;
 
 import static com.aws.jverify.JVerify.*;
