+++ verifier/src/test/java/com/aws/jverify/verifier/tests/interfaces/InterfacesVerification.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.tests;
+package com.aws.jverify.verifier.tests.interfaces;
 
 import com.aws.jverify.*;
 import com.aws.jverify.testengine.JVerifyTest;
