+++ verifier/src/test/java/com/aws/jverify/verifier/tests/nestedClasses/NestedStaticClass.java
@@ -1,4 +1,4 @@
-package com.aws.jverify.verifier.tests;
+package com.aws.jverify.verifier.tests.nestedClasses;
 
 import com.aws.jverify.Nullable;
 import com.aws.jverify.testengine.JVerifyTest;
@@ -6,9 +6,10 @@ import com.aws.jverify.testengine.JVerifyTest;
 import static com.aws.jverify.JVerify.check;
 
 @JVerifyTest(exitCode = 4, dafnyVerified = 0, dafnyErrors = 1)
-public class NestedClass {
-    @Nullable Nestee nestee;
-    public class Nestee {
+public class NestedStaticClass {
+    @Nullable
+    StaticNestee nestee;
+    public static class StaticNestee {
         void checkFalse() {
             check(false);
 //          ^^^^^^^^^^^^ Error: assertion might not hold
@@ -16,7 +17,8 @@ public class NestedClass {
     }
 }
 
-class Nestee {
-    @Nullable Nestee fakeNestee;
-    @Nullable NestedClass.Nestee nestee;
+class StaticNestee {
+    @Nullable
+    StaticNestee fakeNestee;
+    NestedStaticClass.@Nullable StaticNestee nestee;
 }
